feat(api): move to package
Improve the naming of the package.
