+++ api/module.go
@@ -1,4 +1,4 @@
-package v1
+package api
 
 import (
 	"go.uber.org/fx"
