+++ health/module.go
@@ -2,12 +2,11 @@ package health
 
 import (
 	"github.com/alexfalkowski/go-service/health"
-	"github.com/alexfalkowski/idpd/server/health/transport/http"
 	"go.uber.org/fx"
 )
 
 // Module for fx.
 var Module = fx.Options(
-	fx.Provide(http.NewHealthObserver), fx.Provide(http.NewLivenessObserver), fx.Provide(http.NewReadinessObserver),
+	fx.Provide(NewHealthObserver), fx.Provide(NewLivenessObserver), fx.Provide(NewReadinessObserver),
 	fx.Provide(NewRegistrations), health.Module,
 )
