+++ test/lib/idpd/v1/http.rb
@@ -4,7 +4,7 @@ module Idpd
   module V1
     class HTTP < Nonnative::HTTPClient
       def create_pipeline(params, opts = {})
-        post('/v1/pipeline', params.to_json, opts)
+        post('/pipeline', params.to_json, opts)
       end
     end
   end
